requirements.txt
# Environment markers
# https://stackoverflow.com/questions/29222269
uwsgi; platform_system != "Windows"
pick
ipython
django-ninja
django-extensions
django~=3.0
arrow
PyJWT
django-cors-headers
mysqlclient
# cryptography==37.0.4
cryptography
requests
redis~=4.0
loguru
# django-crontab
# django-rq
# django-print-sql
# hashids
# multiavatar
Pillow
qrcode
# python-magic
pycryptodomex
hashids
